enh : dynamically import Mermaid, YouTube, and TwitterTweetEmbed; remove unused Twitter components
