+++ components/MDX/MDX.tsx
@@ -1,20 +1,7 @@
 import { MDXProvider as CoreMDXProvider } from '@mdx-js/react';
-import mermaid from 'mermaid';
+import dynamic from 'next/dynamic';
 import Link from 'next/link';
-import React, { useEffect, useId, useState } from 'react';
-import {
-  TwitterDMButton,
-  TwitterFollowButton,
-  TwitterHashtagButton,
-  TwitterMentionButton,
-  TwitterMomentShare,
-  TwitterOnAirButton,
-  TwitterShareButton,
-  TwitterTimelineEmbed,
-  TwitterTweetEmbed,
-  TwitterVideoEmbed
-} from 'react-twitter-embed';
-import YouTube from 'react-youtube-embed';
+import React, { useId } from 'react';
 
 import Asyncapi3ChannelComparison from '../Asyncapi3Comparison/Asyncapi3ChannelComparison';
 import Asyncapi3IdAndAddressComparison from '../Asyncapi3Comparison/Asyncapi3IdAndAddressComparison';
@@ -39,148 +26,28 @@ import Profiles from '../Profiles';
 import Remember from '../Remember';
 import Sponsors from '../sponsors/PlatinumSponsors';
 import Warning from '../Warning';
-import { Table, TableBody, TableCell, TableHeader, TableRow, Thead } from './MDXTable';
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHeader,
+  TableRow,
+  Thead,
+} from './MDXTable';
 
-type MermaidTheme = 'light' | 'dark';
+const MermaidDiagram = dynamic(() => import('./MermaidDiagram'), {
+  ssr: false,
+});
 
-const MERMAID_THEME_VARIABLES: Record<MermaidTheme, Record<string, string>> = {
-  light: {
-    primaryColor: '#EDFAFF',
-    primaryBorderColor: '#47BCEE',
-    secondaryColor: '#F4EFFC',
-    secondaryBorderColor: '#875AE2',
-    fontFamily: 'Inter, sans-serif',
-    fontSize: '18px',
-    primaryTextColor: '#242929',
-    tertiaryColor: '#F7F9FA',
-    tertiaryBorderColor: '#BFC6C7',
-    lineColor: '#BFC6C7',
-    mainBkg: '#EDFAFF',
-    secondBkg: '#F4EFFC',
-    tertiaryBkg: '#F7F9FA',
-    clusterBkg: '#F7F9FA',
-    clusterBorder: '#BFC6C7',
-    edgeLabelBackground: '#FFFFFF'
-  },
-  dark: {
-    primaryColor: '#1E293B',
-    primaryBorderColor: '#38BDF8',
-    secondaryColor: '#2E2459',
-    secondaryBorderColor: '#A87EFC',
-    fontFamily: 'Inter, sans-serif',
-    fontSize: '18px',
-    primaryTextColor: '#F8FAFC',
-    tertiaryColor: '#121825',
-    tertiaryBorderColor: '#475569',
-    lineColor: '#94A3B8',
-    mainBkg: '#1E293B',
-    secondBkg: '#2E2459',
-    tertiaryBkg: '#121825',
-    clusterBkg: '#121825',
-    clusterBorder: '#475569',
-    edgeLabelBackground: '#1E293B'
-  }
-};
+const TwitterTweetEmbed = dynamic(
+  () =>
+    import('react-twitter-embed').then((mod) => ({
+      default: mod.TwitterTweetEmbed,
+    })),
+  { ssr: false },
+);
 
-// Cache the theme Mermaid was initialized with across client-side page transitions.
-let initializedMermaidTheme: MermaidTheme | null = null;
-
-/**
- * @description Returns the Mermaid theme that matches the current website theme.
- */
-function getMermaidTheme(): MermaidTheme {
-  if (typeof document === 'undefined') {
-    return 'light';
-  }
-
-  return document.documentElement.classList.contains('dark') ? 'dark' : 'light';
-}
-
-/**
- * @description Initializes the Mermaid library for the selected theme.
- */
-function initializeMermaid(theme: MermaidTheme) {
-  if (initializedMermaidTheme === theme) {
-    return;
-  }
-
-  initializedMermaidTheme = theme;
-  mermaid.initialize({
-    startOnLoad: false,
-    theme: 'base',
-    securityLevel: 'strict',
-    // Keep Mermaid styling fully controlled by MERMAID_THEME_VARIABLES.
-    themeCSS: '',
-    themeVariables: MERMAID_THEME_VARIABLES[theme]
-  });
-}
-
-let currentId = 0;
-
-/**
- * @description Generates a unique identifier.
- * @returns {string} - A unique identifier.
- */
-const uuid = (): string => `mermaid-${(currentId++).toString()}`;
-
-interface MermaidDiagramProps {
-  graph: string;
-}
-
-/**
- * @description This component renders Mermaid diagrams.
- *
- * @param {MermaidDiagramProps} props - The props for the MermaidDiagram component.
- * @param {string} props.graph - The Mermaid graph to render.
- */
-function MermaidDiagram({ graph }: Readonly<MermaidDiagramProps>) {
-  const [svg, setSvg] = useState<string | null>(null);
-  const [theme, setTheme] = useState<MermaidTheme>('light');
-
-  useEffect(() => {
-    setTheme(getMermaidTheme());
-
-    const observer = new MutationObserver(() => {
-      setTheme(getMermaidTheme());
-    });
-
-    observer.observe(document.documentElement, { attributes: true, attributeFilter: ['class'] });
-
-    return () => observer.disconnect();
-  }, []);
-
-  /**
-   * @description Renders the Mermaid diagram.
-   */
-  useEffect(() => {
-    let mounted = true;
-
-    if (graph) {
-      try {
-        initializeMermaid(theme);
-        mermaid.mermaidAPI.render(uuid(), graph.trim(), (svgGraph) => {
-          if (mounted) {
-            setSvg(svgGraph);
-          }
-        });
-      } catch (e) {
-        if (mounted) {
-          setSvg(null);
-        }
-        // eslint-disable-next-line no-console
-        console.error(e);
-      }
-    } else {
-      setSvg(null);
-    }
-
-    return () => {
-      mounted = false;
-    };
-  }, [graph, theme]);
-
-  return <div dangerouslySetInnerHTML={{ __html: svg || '' }} />;
-}
+const YouTube = dynamic(() => import('react-youtube-embed'), { ssr: false });
 
 interface CodeComponentProps {
   children: string;
@@ -197,7 +64,12 @@ interface CodeComponentProps {
  * @param {string} [props.className] - The code block class name.
  * @param {string} [props.metastring] - The code block metastring.
  */
-function CodeComponent({ children, className = '', metastring = '', ...rest }: CodeComponentProps) {
+function CodeComponent({
+  children,
+  className = '',
+  metastring = '',
+  ...rest
+}: CodeComponentProps) {
   let caption;
   const meta = metastring.split(/([\w]+=[\w\d\s\-_:><.]+)/) || [];
 
@@ -210,7 +82,8 @@ function CodeComponent({ children, className = '', metastring = '', ...rest }: C
     }
   });
   const maybeLanguage = className.match(/language-([\w\d\-_]+)/);
-  const language = maybeLanguage && maybeLanguage.length >= 2 ? maybeLanguage[1] : undefined;
+  const language =
+    maybeLanguage && maybeLanguage.length >= 2 ? maybeLanguage[1] : undefined;
 
   if (language === 'mermaid') {
     return <MermaidDiagram graph={children} />;
@@ -260,7 +133,8 @@ function extractText(node: React.ReactNode): string {
   if (node == null || typeof node === 'boolean') return '';
   if (typeof node === 'string' || typeof node === 'number') return String(node);
   if (Array.isArray(node)) return node.map(extractText).join('');
-  if (React.isValidElement(node)) return extractText((node.props as { children?: React.ReactNode }).children);
+  if (React.isValidElement(node))
+    return extractText((node.props as { children?: React.ReactNode }).children);
 
   return '';
 }
@@ -288,9 +162,12 @@ const getMDXComponents = (reactId: string) => ({
       {...props}
       id={props.id || generateSlug(props.children) || `heading-${reactId}`}
       className={`${props.className || ''} my-4 font-heading text-2xl font-semibold tracking-heading dark:text-dark-heading text-gray-900 antialiased`}
-      aria-label={props['aria-label'] ?? (extractText(props.children).trim() ? undefined : 'Section title')}
+      aria-label={
+        props['aria-label'] ??
+        (extractText(props.children).trim() ? undefined : 'Section title')
+      }
     >
-      {props.children || <span className='sr-only'>Section title</span>}
+      {props.children || <span className="sr-only">Section title</span>}
     </h1>
   ),
   h2: (props: React.HTMLProps<HTMLHeadingElement>) => (
@@ -298,9 +175,12 @@ const getMDXComponents = (reactId: string) => ({
       {...props}
       id={props.id || generateSlug(props.children) || `heading-${reactId}`}
       className={`${props.className || ''} mb-4 mt-6 font-heading text-2xl font-semibold tracking-heading  dark:text-dark-heading text-gray-900 antialiased`}
-      aria-label={props['aria-label'] ?? (extractText(props.children).trim() ? undefined : 'Section title')}
+      aria-label={
+        props['aria-label'] ??
+        (extractText(props.children).trim() ? undefined : 'Section title')
+      }
     >
-      {props.children || <span className='sr-only'>Section title</span>}
+      {props.children || <span className="sr-only">Section title</span>}
     </h2>
   ),
   h3: (props: React.HTMLProps<HTMLHeadingElement>) => (
@@ -308,9 +188,12 @@ const getMDXComponents = (reactId: string) => ({
       {...props}
       id={props.id || generateSlug(props.children) || `heading-${reactId}`}
       className={`${props.className || ''} mb-4 mt-6 font-heading text-lg font-medium tracking-heading dark:text-dark-heading text-gray-900 antialiased`}
-      aria-label={props['aria-label'] ?? (extractText(props.children).trim() ? undefined : 'Section title')}
+      aria-label={
+        props['aria-label'] ??
+        (extractText(props.children).trim() ? undefined : 'Section title')
+      }
     >
-      {props.children || <span className='sr-only'>Section title</span>}
+      {props.children || <span className="sr-only">Section title</span>}
     </h3>
   ),
   h4: (props: React.HTMLProps<HTMLHeadingElement>) => (
@@ -318,9 +201,12 @@ const getMDXComponents = (reactId: string) => ({
       {...props}
       id={props.id || generateSlug(props.children) || `heading-${reactId}`}
       className={`${props.className || ''} text-md my-4 font-heading font-medium dark:text-dark-heading text-gray-900 antialiased`}
-      aria-label={props['aria-label'] ?? (extractText(props.children).trim() ? undefined : 'Section title')}
+      aria-label={
+        props['aria-label'] ??
+        (extractText(props.children).trim() ? undefined : 'Section title')
+      }
     >
-      {props.children || <span className='sr-only'>Section title</span>}
+      {props.children || <span className="sr-only">Section title</span>}
     </h4>
   ),
   h5: (props: React.HTMLProps<HTMLHeadingElement>) => (
@@ -328,9 +214,12 @@ const getMDXComponents = (reactId: string) => ({
       {...props}
       id={props.id || generateSlug(props.children) || `heading-${reactId}`}
       className={`${props.className || ''} text-md my-4 font-heading dark:text-dark-heading   font-bold antialiased`}
-      aria-label={props['aria-label'] ?? (extractText(props.children).trim() ? undefined : 'Section title')}
+      aria-label={
+        props['aria-label'] ??
+        (extractText(props.children).trim() ? undefined : 'Section title')
+      }
     >
-      {props.children || <span className='sr-only'>Section title</span>}
+      {props.children || <span className="sr-only">Section title</span>}
     </h5>
   ),
   h6: (props: React.HTMLProps<HTMLHeadingElement>) => (
@@ -338,9 +227,12 @@ const getMDXComponents = (reactId: string) => ({
       {...props}
       id={props.id || generateSlug(props.children) || `heading-${reactId}`}
       className={`${props.className || ''} my-4 font-heading text-sm font-bold uppercase dark:text-dark-heading text-gray-900 antialiased`}
-      aria-label={props['aria-label'] ?? (extractText(props.children).trim() ? undefined : 'Section title')}
+      aria-label={
+        props['aria-label'] ??
+        (extractText(props.children).trim() ? undefined : 'Section title')
+      }
     >
-      {props.children || <span className='sr-only'>Section title</span>}
+      {props.children || <span className="sr-only">Section title</span>}
     </h6>
   ),
   blockquote: (props: React.HTMLProps<HTMLQuoteElement>) => (
@@ -377,7 +269,7 @@ const getMDXComponents = (reactId: string) => ({
     <ol
       {...props}
       className={`${props.className || ''} font-normal my-4 ml-4 list-decimal font-body dark:text-dark-heading text-gray-700 antialiased`}
-      type='1'
+      type="1"
     />
   ),
   li: (props: React.HTMLProps<HTMLLIElement>) => (
@@ -386,11 +278,13 @@ const getMDXComponents = (reactId: string) => ({
       className={`${props.className || ''} my-3 font-body font-regular tracking-tight dark:text-white text-gray-700 antialiased`}
     />
   ),
-  button: Button as React.ComponentType<React.ButtonHTMLAttributes<HTMLButtonElement>>,
+  button: Button as React.ComponentType<
+    React.ButtonHTMLAttributes<HTMLButtonElement>
+  >,
   table: (props: React.HTMLProps<HTMLTableElement>) => (
     <div className={`${props.className || ''} flex flex-col`}>
-      <div className='my-2 overflow-x-auto py-2'>
-        <div className='inline-block min-w-full border-b border-gray-200 dark:border-gray-800 align-middle shadow sm:rounded-lg'>
+      <div className="my-2 overflow-x-auto py-2">
+        <div className="inline-block min-w-full border-b border-gray-200 dark:border-gray-800 align-middle shadow sm:rounded-lg">
           <table {...props} className={`${props.className || ''} w-full`} />
         </div>
       </div>
@@ -403,7 +297,10 @@ const getMDXComponents = (reactId: string) => ({
     />
   ),
   tr: (props: React.HTMLProps<HTMLTableRowElement>) => (
-    <tr {...props} className={`${props.className || ''} bg-white dark:bg-transparent`} />
+    <tr
+      {...props}
+      className={`${props.className || ''} bg-white dark:bg-transparent`}
+    />
   ),
   td: (props: React.HTMLProps<HTMLTableCellElement>) => (
     <td
@@ -411,18 +308,25 @@ const getMDXComponents = (reactId: string) => ({
       className={`${props.className || ''} border-b border-gray-200 dark:border-gray-800 px-6 py-4 text-sm leading-5 tracking-tight text-gray-700 dark:text-gray-300`}
     />
   ),
-  pre: (props: React.HTMLProps<HTMLPreElement>) => CodeComponent((props.children as React.ReactElement)?.props),
+  pre: (props: React.HTMLProps<HTMLPreElement>) =>
+    CodeComponent((props.children as React.ReactElement)?.props),
   code: (props: React.HTMLProps<HTMLElement>) => (
     <code
       {...props}
       className={`${props.className || ''} rounded bg-gray-200 dark:bg-gray-800 px-1 py-0.5 font-mono text-sm text-gray-800 dark:text-gray-300`}
     />
   ),
-  hr: (props: React.HTMLProps<HTMLHRElement>) => <hr {...props} className={`${props.className || ''} my-8`} />,
-  Link: ({ href = '/', children, ...props }: React.HTMLProps<HTMLAnchorElement>) => (
+  hr: (props: React.HTMLProps<HTMLHRElement>) => (
+    <hr {...props} className={`${props.className || ''} my-8`} />
+  ),
+  Link: ({
+    href = '/',
+    children,
+    ...props
+  }: React.HTMLProps<HTMLAnchorElement>) => (
     <Link
       href={href as string}
-      className='border-b border-secondary-400 font-body font-semibold text-gray-900 antialiased transition duration-300 ease-in-out hover:border-secondary-500'
+      className="border-b border-secondary-400 font-body font-semibold text-gray-900 antialiased transition duration-300 ease-in-out hover:border-secondary-500"
       {...props}
     >
       {children}
@@ -435,7 +339,10 @@ const getMDXComponents = (reactId: string) => ({
   Thead,
   Table,
   Dl: ({ className, ...props }: React.HTMLProps<HTMLDListElement>) => (
-    <dl {...props} className={`${className || ''} my-4 font-body antialiased`} />
+    <dl
+      {...props}
+      className={`${className || ''} my-4 font-body antialiased`}
+    />
   ),
   Dt: ({ className, ...props }: React.HTMLProps<HTMLElement>) => (
     <dt
@@ -471,18 +378,9 @@ const getMDXComponents = (reactId: string) => ({
   DocsCards,
   GeneratorInstallation,
   NewsletterSubscribe,
-  TwitterTimelineEmbed,
-  TwitterShareButton,
-  TwitterFollowButton,
-  TwitterHashtagButton,
-  TwitterMentionButton,
   TwitterTweetEmbed,
-  TwitterMomentShare,
-  TwitterDMButton,
-  TwitterVideoEmbed,
-  TwitterOnAirButton,
   Profiles,
-  Visualizer
+  Visualizer,
 });
 
 export const mdxComponents = getMDXComponents;
@@ -500,5 +398,9 @@ interface MDXProviderProps {
 export function MDXProvider({ children }: MDXProviderProps) {
   const reactId = useId();
 
-  return <CoreMDXProvider components={mdxComponents(reactId)}>{children}</CoreMDXProvider>;
+  return (
+    <CoreMDXProvider components={mdxComponents(reactId)}>
+      {children}
+    </CoreMDXProvider>
+  );
 }
